+++ src/sources/custom_line_source.cpp
@@ -7,7 +7,7 @@ using namespace std;
 namespace qbox {
 
     custom_line_source::custom_line_source(const surface &surf, double (*time_func)(double)):
-        p1(surf.a.cast<int>()), p2(surf.b.cast<int>()), time_func(time_func) {};
+        p1(surf.a), p2(surf.b), time_func(time_func) {};
 
     void custom_line_source::pulse() {
     }
